moved  publish subscription code in to pubSub.js and access it in server.js through router
